Documentation and code structure consistency
Header guards are now consistent between different header files, with
the general prefix of SELKIELogger<lib>_<file>

Doxygen comments are generally split with the brief comment in the
header file and detailed comments (including parameter and return type
descriptions) in the source files. The obvious exception being typedefs,
structs and defines that only exist in the header files. Doxygen groups
remain in the header files for now, but will probably move to be
documented in separate files later.