Move to more structured approach to source types
Information sources need to define a set of callbacks for startup,
logging and shutdown, as well as a parameters structure containing the
information about that source.

e.g. GPS: Port name, baud rates, handle

These parameter structures will then be passed to each when the
functions are run from main() or via pthreads. This also means that
there is no expectation that only a single MP or GPS source will be
used, although the code to support this at run time is not in place.