Add support for real date/time conversion
If the primary clock source defines an Epoch field (as the default
does), this can be used to convert the arbitrary timestamps to clock
time based on offsets between epoch messages.

Under this implementation, conversion time is increased significantly
(from 4 minutes to 16 minutes on sample data) and remains inactive by
default.